+++ core/src/commonMain/kotlin/visiter/NavigationChainVisitingCallback.kt
@@ -1,0 +1,5 @@
+package dev.inmo.navigation.core.visiter
+
+import dev.inmo.navigation.core.NavigationChain
+
+typealias NavigationChainVisitingCallback<Base> = (NavigationChain<Base>) -> Unit