+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * `NavigationChain` and `NavigationNode` got `findChain` extension
 * `NavigationChain` and `NavigationNode` got `findNode` extension
 * On restoring of hierarchy it is possible to remove the chains which do not fit restoring one
+* Add opportunity to walk across whole navigation tree
 
 ## 0.0.15